+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.16] - 2019-08-18
+### Added
+- Added Array => `deep_dup`
+- Added Hash => `deep_dup`
+- Added Integer => `combinatorial`
+- Added Numeric => `round_down`
+- Added Object => `deep_dup`
+### Changed
+- Changed how integer factorial works
+
 ## [1.0.15] - 2019-08-17
 ### Changed
 - Added config check for each monkey patch